refactors lowercase function to its own module and lowers case of variable names in questions
